Move commands folder into src
