+++ src/commands/loot/index.js
